Added - sunset/sunrise parsing from Yandex. Changed zero symblol for Font11. Added hysteresis for light sensr. Date position changed. Added minus symbol for main temperature value. Added exit from any mode by long pressing button.

+++ main/light_sensor.h
@@ -11,9 +11,9 @@
 void light_sensor_init(void);
 void light_sensor_handling(void);
 uint16_t  light_sensor_get_volt(void);
-uint32_t light_lensor_get_lux(uint32_t volt_mv);
+float light_lensor_get_lux(uint32_t volt_mv);
 void dimming_set_common_brightness(float percent);
-uint32_t light_lensor_get_cur_lux(void);
+float light_lensor_get_cur_lux(void);
 void dimming_auto_controlling(void);
 void dimming_change_auto_controlling(bool is_enabled);
 bool light_sensor_is_night(void);
